Replaced the _file variable to a _redirection variable of the ARequestType class
We now only sets this variable when the results is a redirection.
Before it was also used to store the error pages builded or the status
text.
Those behaviour are removeed.
Status text is determined with the status code
The error page will be builded after with the code and text