Changed the parameters of the flowbuffer
It now takes 3 arguments : T, const void *, size_t
T can be anything.
It can also takes a write and read custom function, if none is entered,
the defaults one are write and read.
For simple fd, we can just pass the fd as T and it works, it only works
with body.
For default fd, we just use write because the fds must have their
options set through fcntl : non blocking for  example.
Added a signal handler for sig pipe. It ignores sig pipe