Changed the ARequestType fd into a _inFd and a _outFd
Two fds because they can be one to write the request content into, and
one to write into the response.
For ewample, for the cgi, there is two : the input and the output of the
process.
For the put there is one : The newly created file to write into
For the get there is one : The file whose content will be sent in the
response.
Added a getter for both.
Added a function to close them and set itt to -1. One for each.
If the two fds are the same, when one is reset, the other one is reseted
too.