Fixeds
Fixe a crash in addFdToListerners that freed the FdData, and the parents
also freed it.
Fixed the _epfd being set to the result of checkError
Added a dash prefix to the statusLine variable.
Removed getInFd and getOutFd.
Added a _bodydestFd into the Request class : now the body class doesn't
have the responsability to close it.
The Execution fo the request handler is now much clearer
addeed a missing checkError call