Added the creation of the Response body -> lot of small changes
Removed the isBlocking variable of the body, it is now in the Request or
Response/RawResponse.
The FdData _fd is now const.
Added a _outSize variable in the ARequestType. It is not set for now !!
Added sizeTToString function in utils/conversion.cpp.
The ABody now has the write fd, not the fd returned by the request, this
fd  is in the response or request class.