Applied the Status class to the code.
The _statuses variable is now a static member of the Status class,
removed it form ARequestType.
The RawResponse doesn't have to convert the cod eto string as the status
class already has it.
Added a function to check if the Status code is a Redirection, error,
client error ... It works with bit operations.