added the moveContentToStart to the flow buffer and fixed the redirecctFdContentToBuffer.
Now when reading, if there is enough characters already written from the
buffer, it will move everything insde at the start. It allows the
FlowBuffer to call read less times and prevents the numCharsWritten to
slowly grow to the end of the buffer.
Also, now, if we have read in this function, it doesn't returns
FLOW_BUFFER_FULL.
It returns FLOW_BUFFER_FULL if it can't read because the buffer is full