Fixed errors in the FlowBuffer.
Fixed misunderstanding of the _contentLength variable.
Now the redirectContent doesn't convert a read FLOW_BUFFER_FULL to
FLOW_MORE when write returns 0. Because that means taht the flowbuffer
is blocked : can't read, nor write. This can appends for example if the
chunk size take the full buffer, we can't read anymore, and the
ChunkedRequest can't write because the chunk size isn't finished.