Refactor the ChunkedBody writeToFd function.
This function now has a single loop.
It also has a static const array of methods, each methods corresponding
to a state.
The while loop get the function with the _state as index and execute it.
writeToFd now call writeChunkedRequestToFd, it is just a wrapper that
get the begin and end parameters.