Fixed the SizedBody
I replaced numCharsToWrite by bufferCapacity in the write function.
I don't return a -1 if the written isn't equal to the number of byte
passed to write because :
If it has been stopped by a signal, the program will crash or the signal
handler will quit the program.
If the computer doesn't have room anymore, the next call will return -1.