Added the beginning of the CgiResponse.
The CgiResponse should : read the headers, when it is done, create the
response firstPart, send the firstPart, and then send the body.
If the cgi returns a bad response, we must set
HTTP_INTERNAL_SERVER_ERROR, and in the best world set the errorPage.
If the body is CgiResponse, the RawResponse should not set the firstPart
itself.