Chunked request (#23)
* added the chunked body class : header, file, canonical form, config. Adde a msggin constructor in the SizedBody

* when reading into the buffer, if it is full : if there is no chars written, return flow buffer full, otherwise, move teh buffer content to the start of the content

* addeed more structuration to the chunked request class

* fixed the ChunkedBody constructor. Added the sum of the the writtenCharacter and the test of error

* The chunked body now use iterators (char *). Added the writeChunkedBodyData code but it doesn t check for the last \n\r for now

* aAdded a function to read the endline of the chunk size. Added a const static string endLine to the ChunkedBody

* Put the state tests in the writeToFd function and added a while loop, We are reading the chunkds while thee state is not CHUNKED_TRAILER or CHUNKED_DONE. When returning a 0, it means there is not enough to read, so we return, if we get a -1 that means there is an error, we also returns but also set the body to finisheed, added the response code and set the tstate

* remove a useless statis variable in the writeToFd file

* added the readTrailers function. Fixed the return of writeToFd when written was to 0 : return totalWritten instead of written. Changed the input fo the request parser : they now take iterators

* applied the change of the flowbuffer getline to functions using it. It now compiles

* renamed the bufferLength to contentLength, _bufferLength to _contentLength and getBufferLength to getContentLength

* Fixed the merge

* Added a _status attribute to the body.

It allows the caller of the body to know why the body has finished.
Set to HTTP_OK if there was no error, otherwise the HTTP error status
code.
The setFinished method now has a status parameter.

* WE now check the status of the body after writing it.

If the status isn't HTTP_OK, we overwrite the previous one.

* Added a strToLOngBase and stringToLOngBase functions

It replaces the stringToSizeT and parseProtocolNumber functions.
It will be useful for the ChunkedRequest too.
It is documented.

* Fixed errors in the FlowBuffer.

Fixed misunderstanding of the _contentLength variable.
Now the redirectContent doesn't convert a read FLOW_BUFFER_FULL to
FLOW_MORE when write returns 0. Because that means taht the flowbuffer
is blocked : can't read, nor write. This can appends for example if the
chunk size take the full buffer, we can't read anymore, and the
ChunkedRequest can't write because the chunk size isn't finished.

* Removed an impossible edge case protection

* Fixed some compilations error

* Changed the ChunkekBody to fit the new methods

It compiles.
It now uses strToLongBase.
Fixed multiple errors.
Removed the response attribute.

* We now handle the SIGTERM signal (for docker)

* Moved the request printing, it now prints when we reset the request

* Fixed an error: the value was put in lowercase instead of the key for the headers.

* Removed duplicates comments in the FlowBuffer class

* Fixed the operator<< overloads.

They need to redirect into the o parameter, and they shouldn't use
std::endl or any function flushing the stream.

* Refactor the ChunkedBody writeToFd function.

This function now has a single loop.
It also has a static const array of methods, each methods corresponding
to a state.
The while loop get the function with the _state as index and execute it.
writeToFd now call writeChunkedRequestToFd, it is just a wrapper that
get the begin and end parameters.

* The ChunkedBody now check if the total written size is superior to a maxSize.

It could overflow for now, I will check every overflow situations soon.

* The setBodyFromHeaders method now also set the chunked request

* Protected the chunked body addition from overflow

* Added th writeOrIgnore method to ABody.

It writes into the fd or ignore the buffer if the fd is set to -1.
getFd is not needed anymore as the child will use it, it is removed.
SizedBody and ChunkedBody call this method.

* Fixed the SizedBody

I replaced numCharsToWrite by bufferCapacity in the write function.
I don't return a -1 if the written isn't equal to the number of byte
passed to write because :
If it has been stopped by a signal, the program will crash or the signal
handler will quit the program.
If the computer doesn't have room anymore, the next call will return -1.

* Fixed a bug when the trailers state would never happend.

It acts the same as the SizedBody for the numCharsWritten,
numCharsToWrite ...

* Added a forgotten constructor in the ChunkedRequest initialisation list

* Body fixes :

After writing a body, we now check first if the body is finished before
checking if the flow buffer returned FLOW_ERROR.
Removed _numCharsWritten, we now decrement _chunkSize.
Fixed a misunderstanding of the format of chunked request

* Added a test for the chunked request -> nothing makes sense

* Updated the makeawRequest script

It now parsed the return chunk by chunk and stop when the response is
finished. We doesn't have to wait for the server to close the
connection.

* Removed unecessary files

* The Makefile now logs the nginx and webserv output.

* Added real chunk data test

* Renamed a parameter from get ro req in the addRoot function

* Restarting webserv also restart the logs

* Fixed the request_method behaviour.

Added Put to the request_method parsing.
Added routes for the put and delete to allow these methods.
The defaults allowed methods are now GET and POST (if no request_method
is set in a route).
If we are nott in a route, PUT and DELETE are not allowed.

* Fixed the chunked tests

Divided the put tests into put and chunked. Added th chunked inti and
clean bash scripts.
Updated the configurations of webserv and nginx for them to accept PUT.

* Changed the gitignore to ignore logs

* Added tests to put that doesnt pass

* Added ranodm chunked request generation

* Added a print variable to show or hide output.

* Replace print by printOK, it determine if OK verify should print, KO always print

* Cleaned the .mjs tests

* In ChunkedRequest, when the body is too big, returns HTTP_CONTENT_TOO_LARGe instead of HTTP_BAD_REQUEST

* Fixed the trailers of the ChunkedBody

It missed the \r, the parseHeader verify that it is present.

* Added tests and upgrade the sendGoodChunkedRequest

This functon now takes the list of headers, chunks and trailers and
create the response.

* Added more tests

* Upgrade the mjs tests

* Added more rules to restart nginx or both servers

* Changed a status text to match nginx's

* Creating the response even on net error, it could cause problemes afterward.

* Added and fix request to test the max size : nginx buffer every request before sending it to a file

* Removed th trailer parsingg in the ChunkedRequest as it is optionnnal

* Added more tests, all passes

* Added  comments

---------

Co-authored-by: brappo <brappo@PCTOPACHAT>
Co-authored-by: Benjamin RAPPO <brappo@desktop-debian-home.home>