FileFd changes.
The constructors now takes a const char* instead of a const
std::string&.
The getTemporaryFile now should only be called if we want to create the
temporary file. It puts the results in the tmpnam parameters without
checking.
The getTemporaryFile now remove it just after creation so we don't need
to remove it later. It is automatically removed when no fd reference it.