EPOLLRDHUP and EPOLLERR doesn't deactivate the ConnectedSocket.
For the same reason as the CgiIn, it is not an error if it has written
enough to handle the request.
On the other end, we don't know if there is not enough input to handler
the request.
WE can't do the same way as the CgiIn because there is no class to call
our callback with no events.