Added setFd and a new constructor to the AFdData.
The setFd changes the fd, apply the flags to the new fd, and add it to
the epoll interest list if the evetns are not 0.
The new constructor takes the ePollHandler but not the fd.
It is usefull when you don't know the blocking fd at the class creation.
We can call the constructor, the fd will be set to -1, get the fd, and
then call setFd.
Added a bool _addedToEPoll, for the destructor to know or nott if we
should remove it from the interest list.