RequestContext now only usable by ARequestType + small setArgv and setEnv changes
Th RequestContext attributes are now private, but the class is friend
with the ARequestTYpe. That means that only the ARequestType can use its
attributes.
setEnv became a ARequestType method as it uses the requestContext.
readRequest now takes a Response as it can no longer takes the
RequestContext's response attribute.
The setEnv and setArgv now throws and don't deallocate env and argv
themselves.