ASocketData changes :
CgiIn and cgiOut are now ASocketData. That means that every blocking fds
are in the EPollHandler list, it will be useful for the timeout.
The ASocketData now has the ePollHandler.
The ASocketData doesn't destroy themselves anymore, they add themselves
in the _socketsToRemoveList, and after all the callback, they are
removed.
It prevents use after free.
The removeFromEPoll now set _active to false + add the fd to the remove
list.
So if they are called, they don't do nothing be cause _active is set to
false.
All ASocketData callback check active before doing anything.