Fixed an error in the FlowBuffer.
If the writeState was done, meaning everything in the buffer was
written, and the readState was FLOW_BUFFER_FULL, meaning it can't write
into the buffer because there is no more room, we returned
FLOW_BUFFER_FULL.
It should be FLOW_MORE, because now, there is more room to write.
FIxed that.