Divided the CgiIn class into two : CgiInChunked and CgiInSized. Also added a temporary buffer to the CgiInChunked.
Even though the number of class is increased, the flow is clearer.
The temporary flow buffer was needed, because if the client sent
multiple request, we would have wrote the request into the cgi.
Having two different classes allows us to only have the buffer (annd all
the other variables if needed)